requirements.txt
# ChroLens_AutoFlow 依賴套件
# 自動生成於 2026-01-19

Pillow
PyQt6
numpy
opencv-python
packaging
requests
webbrowser
